requirements.txt
Django==2.1.5
django-tinymce==2.8.0
pytz==2018.9
